Set up GoodJob

+++ db/views/race_records_v01.sql
@@ -1,68 +1,68 @@
 SELECT
     rr.horse_id,
-    DATE_PART('Year', r.date) AS year,
+    (DATE_PART('Year', r.date))::integer AS year,
     ts.surface,
-    count(rr.id) AS starts,
+    count(rr.id)::integer AS starts,
     SUM(
             CASE
                 WHEN r.race_type = 'stakes' THEN 1
                 ELSE 0
                 END
-    ) AS stakes_starts,
+    )::integer AS stakes_starts,
     SUM(
             CASE finish_position
                 WHEN 1 THEN 1
                 ELSE 0
                 END
-    ) AS wins,
+    )::integer AS wins,
     SUM(
             CASE
                 WHEN r.race_type = 'stakes'
                     AND finish_position = 1 THEN 1
                 ELSE 0
                 END
-    ) AS stakes_wins,
+    )::integer AS stakes_wins,
     SUM(
             CASE finish_position
                 WHEN 2 THEN 1
                 ELSE 0
                 END
-    ) AS seconds,
+    )::integer AS seconds,
     SUM(
             CASE
                 WHEN r.race_type = 'stakes'
                     AND finish_position = 2 THEN 1
                 ELSE 0
                 END
-    ) AS stakes_seconds,
+    )::integer AS stakes_seconds,
     SUM(
             CASE finish_position
                 WHEN 3 THEN 1
                 ELSE 0
                 END
-    ) AS thirds,
+    )::integer AS thirds,
     SUM(
             CASE
                 WHEN r.race_type = 'stakes'
                     AND finish_position = 3 THEN 1
                 ELSE 0
                 END
-    ) AS stakes_thirds,
+    )::integer AS stakes_thirds,
     SUM(
             CASE finish_position
                 WHEN 4 THEN 1
                 ELSE 0
                 END
-    ) AS fourths,
+    )::integer AS fourths,
     SUM(
             CASE
                 WHEN r.race_type = 'stakes'
                     AND finish_position = 4 THEN 1
                 ELSE 0
                 END
-    ) AS stakes_fourths,
-    SUM(rr.earnings) AS earnings,
-    SUM(rr.points) AS points
+    )::integer AS stakes_fourths,
+    SUM(rr.earnings)::bigint AS earnings,
+    SUM(rr.points)::integer AS points
 FROM
     race_result_horses rr
         LEFT JOIN race_results r ON rr.race_id = r.id
